Revert "修改行号类型"
This reverts commit e294f033

+++ CommonKits/src/main/java/io/github/xuyao5/dkl/common/standard/StandardFileLine.java
@@ -13,6 +13,6 @@ import java.io.Serializable;
 @Data(staticConstructor = "of")
 public final class StandardFileLine implements Serializable {
 
-    private long lineNo;//行号
+    private int lineNo;//行号
     private String lineRecord;//行记录
 }
